+++ test/integration/originRedirect.test.js
@@ -22,7 +22,8 @@ describe('passwordless', function() {
 			passwordless.init(store);
 			passwordless.addDelivery(mocks.deliveryMockSend());
 
-			app.use(bodyParser());
+			app.use(bodyParser.json());
+			app.use(bodyParser.urlencoded({extended: false}));
 			app.use(cookieParser());
 			app.use(expressSession({ secret: '42', resave: false, saveUninitialized:false }));