+++ test/integration/standard.test.js
@@ -21,7 +21,8 @@ describe('passwordless', function() {
 			passwordless.init(new TokenStoreMock({integration:true}));
 			passwordless.addDelivery(mocks.deliveryMockSend());
 
-			app.use(bodyParser());
+			app.use(bodyParser.json());
+			app.use(bodyParser.urlencoded({extended: false}));
 			app.use(cookieParser());
 			app.use(expressSession({ secret: '42', resave: false, saveUninitialized:false }));