+++ test/integration/two-step-auth.test.js
@@ -20,8 +20,9 @@ describe('passwordless', function() {
 			var passwordless = new Passwordless();
 			passwordless.init(new TokenStoreMock({integration:true}));
 			passwordless.addDelivery(mocks.deliveryMockSend());
-
-			app.use(bodyParser());
+			
+			app.use(bodyParser.json());
+			app.use(bodyParser.urlencoded({extended: false}));
 			app.use(cookieParser());
 			app.use(expressSession({ secret: '42', resave: false, saveUninitialized:false }));